% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Occupation.R
\name{Occupation}
\alias{Occupation}
\title{R6 Class pour donnees occupations des parkings}
\description{
va gérer la routine API / clean / plot / table / download
}
\examples{

## ------------------------------------------------
## Method `Occupation$download_data`
## ------------------------------------------------

\dontrun{
parc_relais <- Occupation(rangeStart = Sys.Date() - 2, 
rangeEnd = Sys.Date() - 1, localisation_parking = NA, parc_relais = TRUE)
parc_relais$download_data()
parc_relais$data_xtradata }

## ------------------------------------------------
## Method `Occupation$clean_output`
## ------------------------------------------------

\dontrun{ clean_output()
}

## ------------------------------------------------
## Method `Occupation$mean_by_some_time_unit`
## ------------------------------------------------

\dontrun{ temporal_aggregate("day")
} 
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{rangeStart}}{Debut de la periode d'observation}

\item{\code{rangeEnd}}{Fin de la periode d'observation}

\item{\code{localisation_parking}}{Secteur de localisation du parking (hypercentre, centre, peripherie, NA pour les parc relais)}

\item{\code{parc_relais}}{Parc relais ou non (boolean)}

\item{\code{data_xtradata}}{Données issues de l'appel au WS via la fonction download_data}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{Occupation$new()}}
\item \href{#method-download_data}{\code{Occupation$download_data()}}
\item \href{#method-clean_output}{\code{Occupation$clean_output()}}
\item \href{#method-mean_by_some_time_unit}{\code{Occupation$mean_by_some_time_unit()}}
\item \href{#method-clone}{\code{Occupation$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Create a new occupation object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Occupation$new(
  rangeStart,
  rangeEnd,
  localisation_parking,
  parc_relais,
  data_xtradata
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{rangeStart}}{rangeStart}

\item{\code{rangeEnd}}{rangeEnd.}

\item{\code{localisation_parking}}{localisation_parking}

\item{\code{parc_relais}}{parc_relais}

\item{\code{data_xtradata}}{data_xtradata}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new `Occupation` object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-download_data"></a>}}
\if{latex}{\out{\hypertarget{method-download_data}{}}}
\subsection{Method \code{download_data()}}{
Interroge le WS aggregate
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Occupation$download_data(rangeStep)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{rangeStep}}{rangeStep xtradata aggregate}
}
\if{html}{\out{</div>}}
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{\dontrun{
parc_relais <- Occupation(rangeStart = Sys.Date() - 2, 
rangeEnd = Sys.Date() - 1, localisation_parking = NA, parc_relais = TRUE)
parc_relais$download_data()
parc_relais$data_xtradata }
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clean_output"></a>}}
\if{latex}{\out{\hypertarget{method-clean_output}{}}}
\subsection{Method \code{clean_output()}}{
Nettoyage de la sortie xtradata
(application de lubridate et calcul du taux d'occup)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Occupation$clean_output()}\if{html}{\out{</div>}}
}

\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{\dontrun{ clean_output()
}
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-mean_by_some_time_unit"></a>}}
\if{latex}{\out{\hypertarget{method-mean_by_some_time_unit}{}}}
\subsection{Method \code{mean_by_some_time_unit()}}{
Aggregation des données selon une fenetre temporelle
(application de la fonction summarise_by_time de timetk)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Occupation$mean_by_some_time_unit(time_unit)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{time_unit}}{pas d'aggreg (voir timetk)}
}
\if{html}{\out{</div>}}
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{\dontrun{ temporal_aggregate("day")
} 
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Occupation$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
