% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_xtradata_request_parameters.R
\name{occupation_compute_xtradata_request_parameters}
\alias{occupation_compute_xtradata_request_parameters}
\title{Calcule les parametres rangeStart, rangeEnd et rangeStep pour xtradata}
\usage{
occupation_compute_xtradata_request_parameters(
  selected_timestep,
  selected_date
)
}
\arguments{
\item{selected_timestep}{pas de temps choisi par l'user (jour, semaine, mois, annee)}

\item{selected_date}{date selectionnee}
}
\value{
une liste de parametres
}
\description{
Calcule les parametres rangeStart, rangeEnd et rangeStep pour xtradata
}
\examples{
occupation_compute_xtradata_request_parameters("Jour", "2021-04-01")
}
