% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parkings.R
\docType{data}
\name{parkings}
\alias{parkings}
\title{Liste des parkings explorés}
\format{
Data frame avec 3 colonnes
\describe{
  \item{localisation_parking}{Secteur du parking - char}
  \item{ident}{Identifiant du parking - char}
  \item{parc relais}{S'agit il d'un parc relais - lgcl}
}
}
\usage{
parkings
}
\description{
Liste des parkings explorés
}
\keyword{datasets}
