% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R6ParkingsStats.R
\name{ParkingsStats}
\alias{ParkingsStats}
\title{R6 Super Class pour donnees occupations et saturation des parkings}
\description{
utilisee ensuite par classe Occupation et Saturation
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{rangeStart}}{Debut de la periode d'observation}

\item{\code{rangeEnd}}{Fin de la periode d'observation}

\item{\code{rangeStep}}{Pas d'aggregation pour requete xtradata}

\item{\code{aggregation_unit}}{Unite de temps d'aggregation des donnees xtradata (notamment pour l'affichage dans les graphes)}

\item{\code{plageHoraire}}{plage horaire des donnees à recup}

\item{\code{parkings_list}}{liste des parkings analyses}

\item{\code{data_xtradata}}{Données issues de l'appel au WS via la fonction download_data}

\item{\code{cleaned_data}}{Données nettoyées}

\item{\code{download_data_memoise}}{version optimisee (avec cache) de la fonction de telechargement sur xtradata}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{ParkingsStats$new()}}
\item \href{#method-download_data}{\code{ParkingsStats$download_data()}}
\item \href{#method-clean_output}{\code{ParkingsStats$clean_output()}}
\item \href{#method-clone}{\code{ParkingsStats$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Create a new occupation object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ParkingsStats$new(
  rangeStart = NULL,
  rangeEnd = NULL,
  rangeStep = NULL,
  aggregation_unit = NULL,
  plageHoraire = NULL,
  parkings_list = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{rangeStart}}{rangeStart}

\item{\code{rangeEnd}}{rangeEnd}

\item{\code{rangeStep}}{rangeStep}

\item{\code{aggregation_unit}}{aggregation_unit}

\item{\code{plageHoraire}}{plageHoraire}

\item{\code{parkings_list}}{liste des parkings analyses}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new `Occupation` object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-download_data"></a>}}
\if{latex}{\out{\hypertarget{method-download_data}{}}}
\subsection{Method \code{download_data()}}{
Interroge le WS aggregate
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ParkingsStats$download_data(
  rangeStart,
  rangeEnd,
  rangeStep,
  plageHoraire,
  parkings_list
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{rangeStart}}{rangeStart xtradata aggregate}

\item{\code{rangeEnd}}{rangeEnd xtradata aggregate}

\item{\code{rangeStep}}{rangeStep xtradata aggregate}

\item{\code{plageHoraire}}{plage horaire d'interet pour les donnees (filtre sur xtradata)}

\item{\code{parkings_list}}{liste des parkings analyses}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clean_output"></a>}}
\if{latex}{\out{\hypertarget{method-clean_output}{}}}
\subsection{Method \code{clean_output()}}{
Nettoyage de la sortie xtradata
(application de lubridate et calcul du taux d'occup)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ParkingsStats$clean_output(parkings_list)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{parkings_list}}{liste des parkings (permet de recup les noms des pkgs)}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ParkingsStats$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
