% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Saturation.R
\name{Saturation}
\alias{Saturation}
\title{R6 Class pour donnees occupations des parkings}
\description{
va gérer la routine API / clean / plot / table / download
}
\examples{

## ------------------------------------------------
## Method `Saturation$download_data`
## ------------------------------------------------

\dontrun{
parc_relais <- Occupation(rangeStart = Sys.Date() - 2, 
rangeEnd = Sys.Date() - 1, localisation_parking = NA, parc_relais = TRUE)
parc_relais$download_data()
parc_relais$data_xtradata }

## ------------------------------------------------
## Method `Saturation$clean_output`
## ------------------------------------------------

\dontrun{ clean_output()
}

## ------------------------------------------------
## Method `Saturation$filter_full_capacity_parkings`
## ------------------------------------------------

\dontrun{ temporal_aggregate("day")
} 

## ------------------------------------------------
## Method `Saturation$calendar_heatmap`
## ------------------------------------------------

\dontrun{ temporal_aggregate("day")
} 
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{rangeStart}}{Debut de la periode d'observation}

\item{\code{rangeEnd}}{Fin de la periode d'observation}

\item{\code{localisation_parking}}{Secteur de localisation du parking (hypercentre, centre, peripherie, NA pour les parc relais)}

\item{\code{parc_relais}}{Parc relais ou non (boolean)}

\item{\code{data_xtradata}}{Données issues de l'appel au WS via la fonction download_data}

\item{\code{parkings_satures}}{contient les parkings satures selon des criteres de % 
d'occupation en nb d'heures par jour et nb de jours par semaine}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{Saturation$new()}}
\item \href{#method-download_data}{\code{Saturation$download_data()}}
\item \href{#method-clean_output}{\code{Saturation$clean_output()}}
\item \href{#method-filter_full_capacity_parkings}{\code{Saturation$filter_full_capacity_parkings()}}
\item \href{#method-calendar_heatmap}{\code{Saturation$calendar_heatmap()}}
\item \href{#method-clone}{\code{Saturation$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Create a new occupation object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Saturation$new(rangeStart, rangeEnd, localisation_parking, parc_relais)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{rangeStart}}{rangeStart}

\item{\code{rangeEnd}}{rangeEnd.}

\item{\code{localisation_parking}}{localisation_parking}

\item{\code{parc_relais}}{parc_relais}

\item{\code{data_xtradata}}{data_xtradata}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new `Occupation` object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-download_data"></a>}}
\if{latex}{\out{\hypertarget{method-download_data}{}}}
\subsection{Method \code{download_data()}}{
Interroge le WS aggregate
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Saturation$download_data(rangeStep)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{rangeStep}}{rangeStep xtradata aggregate}
}
\if{html}{\out{</div>}}
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{\dontrun{
parc_relais <- Occupation(rangeStart = Sys.Date() - 2, 
rangeEnd = Sys.Date() - 1, localisation_parking = NA, parc_relais = TRUE)
parc_relais$download_data()
parc_relais$data_xtradata }
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clean_output"></a>}}
\if{latex}{\out{\hypertarget{method-clean_output}{}}}
\subsection{Method \code{clean_output()}}{
Nettoyage de la sortie xtradata
(application de lubridate et calcul du taux d'occup)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Saturation$clean_output()}\if{html}{\out{</div>}}
}

\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{\dontrun{ clean_output()
}
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-filter_full_capacity_parkings"></a>}}
\if{latex}{\out{\hypertarget{method-filter_full_capacity_parkings}{}}}
\subsection{Method \code{filter_full_capacity_parkings()}}{
On garde les parkings satures. Càd les parkings avec un taux d'occupation
superieur à seuil_taux_occupation pendant au moins nb_heures_par_jour_satures 
par jour pendant au moins nb_jour_par_semaine_sature par semaine
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Saturation$filter_full_capacity_parkings(
  seuil_taux_occupation = 0.9,
  nb_heures_par_jour_satures = 3,
  nb_jour_par_semaine_sature = 2
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{seuil_taux_occupation}}{seuil pour considerer un parking comme sature}

\item{\code{nb_heures_par_jour_satures}}{seuil de nb d'heures par jour de saturation}

\item{\code{nb_jour_par_semaine_sature}}{seuil de nb de jour par semaine de saturation}
}
\if{html}{\out{</div>}}
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{\dontrun{ temporal_aggregate("day")
} 
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-calendar_heatmap"></a>}}
\if{latex}{\out{\hypertarget{method-calendar_heatmap}{}}}
\subsection{Method \code{calendar_heatmap()}}{
Realise une calendar heatmap des parkings les plus satures
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Saturation$calendar_heatmap()}\if{html}{\out{</div>}}
}

\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{\dontrun{ temporal_aggregate("day")
} 
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Saturation$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
