% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R6Occupation.R
\name{Occupation}
\alias{Occupation}
\title{R6 Class pour donnees occupations des parkings}
\description{
va gérer la routine API / clean / plot / table / download
}
\examples{

## ------------------------------------------------
## Method `Occupation$mean_by_some_time_unit`
## ------------------------------------------------

\dontrun{ temporal_aggregate("day")
} 

## ------------------------------------------------
## Method `Occupation$timeseries_plot`
## ------------------------------------------------

\dontrun{ timeseries_plot(parkings_to_plot = c("A","B"), show_average = TRUE)
} 
}
\section{Super class}{
\code{\link[occupationparkingsapp:ParkingsStats]{occupationparkingsapp::ParkingsStats}} -> \code{Occupation}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-mean_by_some_time_unit}{\code{Occupation$mean_by_some_time_unit()}}
\item \href{#method-timeseries_plot}{\code{Occupation$timeseries_plot()}}
\item \href{#method-clone}{\code{Occupation$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="occupationparkingsapp" data-topic="ParkingsStats" data-id="clean_output">}\href{../../occupationparkingsapp/html/ParkingsStats.html#method-clean_output}{\code{occupationparkingsapp::ParkingsStats$clean_output()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="occupationparkingsapp" data-topic="ParkingsStats" data-id="download_data">}\href{../../occupationparkingsapp/html/ParkingsStats.html#method-download_data}{\code{occupationparkingsapp::ParkingsStats$download_data()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="occupationparkingsapp" data-topic="ParkingsStats" data-id="initialize">}\href{../../occupationparkingsapp/html/ParkingsStats.html#method-initialize}{\code{occupationparkingsapp::ParkingsStats$initialize()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-mean_by_some_time_unit"></a>}}
\if{latex}{\out{\hypertarget{method-mean_by_some_time_unit}{}}}
\subsection{Method \code{mean_by_some_time_unit()}}{
Aggregation des données selon une fenetre temporelle
(application de la fonction summarise_by_time de timetk)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Occupation$mean_by_some_time_unit(time_unit, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{time_unit}}{pas d'aggreg (voir timetk)}

\item{\code{...}}{parametres additionels de floor_date}
}
\if{html}{\out{</div>}}
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{\dontrun{ temporal_aggregate("day")
} 
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-timeseries_plot"></a>}}
\if{latex}{\out{\hypertarget{method-timeseries_plot}{}}}
\subsection{Method \code{timeseries_plot()}}{
Graphe de série temporelle
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Occupation$timeseries_plot(parkings_to_plot, show_average = TRUE)}\if{html}{\out{</div>}}
}

\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{\dontrun{ timeseries_plot(parkings_to_plot = c("A","B"), show_average = TRUE)
} 
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Occupation$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
