% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parkings.R
\docType{data}
\name{parkings}
\alias{parkings}
\title{Parkings analyses}
\format{
A data frame with 49 rows and 3 variables:
\describe{
  \item{localisation_parking}{secteur}
  \item{ident}{identifiant}
  \item{nom}{nom}
  \item{parc_relais}{parc_relais}
}
}
\usage{
parkings
}
\description{
A dataset containing the prices and other attributes of almost 54,000
diamonds.
}
\keyword{datasets}
