% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R6ParkingsStats.R
\name{ParkingsStats}
\alias{ParkingsStats}
\title{R6 Super Class pour donnees occupations et saturation des parkings}
\description{
utilisee ensuite par classe Occupation et Saturation
}
\examples{

## ------------------------------------------------
## Method `ParkingsStats$download_data`
## ------------------------------------------------

\dontrun{}

## ------------------------------------------------
## Method `ParkingsStats$clean_output`
## ------------------------------------------------

\dontrun{ clean_output()
}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{rangeStart}}{Debut de la periode d'observation}

\item{\code{rangeEnd}}{Fin de la periode d'observation}

\item{\code{rangeStep}}{Pas d'aggregation pour requete xtradata}

\item{\code{timeStep}}{Fenetre de temps de restitution des graphes (heure, jour, jour de la semaine, mois)}

\item{\code{plageHoraire}}{plage horaire des donnees à recup}

\item{\code{localisation_parking}}{Secteur de localisation du parking (hypercentre, centre, peripherie, NA pour les parc relais)}

\item{\code{parc_relais}}{Parc relais ou non (boolean)}

\item{\code{data_xtradata}}{Données issues de l'appel au WS via la fonction download_data}

\item{\code{cleaned_data}}{Données nettoyées}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{ParkingsStats$new()}}
\item \href{#method-download_data}{\code{ParkingsStats$download_data()}}
\item \href{#method-clean_output}{\code{ParkingsStats$clean_output()}}
\item \href{#method-clone}{\code{ParkingsStats$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Create a new occupation object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ParkingsStats$new(
  rangeStart,
  rangeEnd,
  rangeStep,
  timeStep,
  plageHoraire,
  localisation_parking,
  parc_relais
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{rangeStart}}{rangeStart}

\item{\code{rangeEnd}}{rangeEnd}

\item{\code{rangeStep}}{rangeStep}

\item{\code{timeStep}}{timeStep}

\item{\code{plageHoraire}}{plageHoraire}

\item{\code{localisation_parking}}{localisation_parking}

\item{\code{parc_relais}}{parc_relais}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new `Occupation` object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-download_data"></a>}}
\if{latex}{\out{\hypertarget{method-download_data}{}}}
\subsection{Method \code{download_data()}}{
Interroge le WS aggregate
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ParkingsStats$download_data(rangeStep)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{rangeStep}}{rangeStep xtradata aggregate}
}
\if{html}{\out{</div>}}
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{\dontrun{}
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clean_output"></a>}}
\if{latex}{\out{\hypertarget{method-clean_output}{}}}
\subsection{Method \code{clean_output()}}{
Nettoyage de la sortie xtradata
(application de lubridate et calcul du taux d'occup)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ParkingsStats$clean_output()}\if{html}{\out{</div>}}
}

\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{\dontrun{ clean_output()
}
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ParkingsStats$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
