% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R6Saturation.R
\name{Saturation}
\alias{Saturation}
\title{R6 Class pour donnees occupations des parkings}
\description{
va gérer la routine API / clean / plot / table / download
}
\examples{

## ------------------------------------------------
## Method `Saturation$filter_full_capacity_parkings`
## ------------------------------------------------

\dontrun{ temporal_aggregate("day")
} 

## ------------------------------------------------
## Method `Saturation$calendar_heatmap`
## ------------------------------------------------

\dontrun{ temporal_aggregate("day")
} 
}
\section{Super class}{
\code{\link[occupationparkingsapp:ParkingsStats]{occupationparkingsapp::ParkingsStats}} -> \code{Saturation}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{parkings_satures}}{contient les parkings satures selon des criteres de % 
d'occupation en nb d'heures par jour et nb de jours par semaine}

\item{\code{data_plot}}{donnees du graphique de saturation}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-filter_full_capacity_parkings}{\code{Saturation$filter_full_capacity_parkings()}}
\item \href{#method-calendar_heatmap}{\code{Saturation$calendar_heatmap()}}
\item \href{#method-clone}{\code{Saturation$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="occupationparkingsapp" data-topic="ParkingsStats" data-id="clean_output">}\href{../../occupationparkingsapp/html/ParkingsStats.html#method-clean_output}{\code{occupationparkingsapp::ParkingsStats$clean_output()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="occupationparkingsapp" data-topic="ParkingsStats" data-id="download_data">}\href{../../occupationparkingsapp/html/ParkingsStats.html#method-download_data}{\code{occupationparkingsapp::ParkingsStats$download_data()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="occupationparkingsapp" data-topic="ParkingsStats" data-id="initialize">}\href{../../occupationparkingsapp/html/ParkingsStats.html#method-initialize}{\code{occupationparkingsapp::ParkingsStats$initialize()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-filter_full_capacity_parkings"></a>}}
\if{latex}{\out{\hypertarget{method-filter_full_capacity_parkings}{}}}
\subsection{Method \code{filter_full_capacity_parkings()}}{
On garde les parkings satures. Càd les parkings avec un taux d'occupation
superieur à seuil_taux_occupation pendant au moins nb_heures_par_jour_satures 
par jour pendant au moins nb_jour_par_semaine_sature par semaine
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Saturation$filter_full_capacity_parkings(
  seuil_taux_occupation = 90,
  nb_heures_par_jour_satures = 3,
  nb_jour_par_semaine_sature = 2
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{seuil_taux_occupation}}{seuil pour considerer un parking comme sature}

\item{\code{nb_heures_par_jour_satures}}{seuil de nb d'heures par jour de saturation}

\item{\code{nb_jour_par_semaine_sature}}{seuil de nb de jour par semaine de saturation}
}
\if{html}{\out{</div>}}
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{\dontrun{ temporal_aggregate("day")
} 
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-calendar_heatmap"></a>}}
\if{latex}{\out{\hypertarget{method-calendar_heatmap}{}}}
\subsection{Method \code{calendar_heatmap()}}{
Realise une calendar heatmap des parkings les plus satures
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Saturation$calendar_heatmap(selected_parking, app_theme)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{selected_parking}}{parkings à afficher}

\item{\code{app_theme}}{theme de l'application (dark ou light)}
}
\if{html}{\out{</div>}}
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{\dontrun{ temporal_aggregate("day")
} 
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Saturation$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
